+++ src/components/common/Card.jsx
@@ -1,4 +1,5 @@
 import { memo } from 'react';
+
 import styled, { css } from 'styled-components';
 
 import PropTypes from 'prop-types';
