+++ src/hooks/useCardNumber.jsx
@@ -1,4 +1,4 @@
-import { useState, useCallback } from 'react';
+import { useCallback, useState } from 'react';
 
 import { ERROR_MESSAGE } from 'constants';
