+++ src/hooks/useSubmit.jsx
@@ -1,4 +1,4 @@
-import { useNavigate } from 'react-router';
+import { useNavigate } from 'react-router-dom';
 import { useCallback } from 'react';
 
 export default function useSubmit(nextPath) {
