+++ src/pages/index.js
@@ -1,0 +1,5 @@
+import AddCard from 'pages/AddCard';
+import AddCardComplete from 'pages/AddCardComplete';
+import CardList from 'pages/CardList';
+
+export { AddCard, AddCardComplete, CardList };
