+++ src/stories/Button.stories.jsx
@@ -1,6 +1,4 @@
-import React from 'react';
-
-import Button from '../components/common/Button';
+import { Button } from 'components';
 
 export default {
   title: 'Example/Button',
