+++ src/stories/Input.stories.jsx
@@ -1,6 +1,4 @@
-import React from 'react';
-
-import Input from '../components/common/Input';
+import { Input } from 'components';
 
 export default {
   title: 'Example/Input',
