+++ src/stories/app.stories.jsx
@@ -1,6 +1,4 @@
-import React from 'react';
-
-import App from '../App';
+import App from 'App';
 
 export default {
   title: 'Example/App',
